requirements.txt
# torch
# torchvision
# numpy
# pillow
# matplotlib
# scikit-learn

torch
torchvision
pillow
numpy
streamlit
